make `Cast` trait have arbitrary outputs
